+++ catalyst/utils/torch.py
@@ -1,4 +1,4 @@
-from typing import Callable, Dict, Iterable, List, Union
+from typing import Any, Callable, Dict, Iterable, List, Mapping, Union
 import collections
 import os
 import re
@@ -10,8 +10,17 @@ import torch.backends
 from torch.backends import cudnn
 
 from catalyst.settings import SETTINGS
-from catalyst.typing import Device, Model, Optimizer
-from catalyst.utils.misc import merge_dicts
+from catalyst.typing import (
+    Device,
+    Model,
+    Optimizer,
+    RunnerCriterion,
+    RunnerModel,
+    RunnerOptimizer,
+    RunnerScheduler,
+)
+from catalyst.utils.distributed import get_nn_from_ddp_module
+from catalyst.utils.misc import maybe_recursive_call, merge_dicts
 
 # TODO: move to global registry with activation functions
 ACTIVATIONS = {  # noqa: WPS407
@@ -231,6 +240,8 @@ def any2device(value, device: Device):
         return {k: any2device(value[k], device) for k in value.dtype.fields.keys()}
     elif isinstance(value, np.ndarray):
         return torch.tensor(value, device=device)
+    elif isinstance(value, nn.Module):
+        return value.to(device)
     return value
 
 
@@ -396,7 +407,7 @@ def get_network_output(net: Model, *input_shapes_args, **input_shapes_kwargs):
     return output_t
 
 
-def detach(tensor: torch.Tensor) -> np.ndarray:
+def detach_tensor(tensor: torch.Tensor) -> np.ndarray:
     """Detach a pytorch tensor from graph and
     convert it to numpy array
 
@@ -431,6 +442,123 @@ def trim_tensors(tensors: Tensor) -> List[torch.Tensor]:
     return tensors
 
 
+def pack_checkpoint(
+    model: RunnerModel = None,
+    criterion: RunnerCriterion = None,
+    optimizer: RunnerOptimizer = None,
+    scheduler: RunnerScheduler = None,
+    **kwargs,
+) -> Dict:
+    """
+    Packs ``model``, ``criterion``, ``optimizer``, ``scheduler``
+    and some extra info ``**kwargs`` to torch-based checkpoint.
+
+    Args:
+        model: torch model
+        criterion: torch criterion
+        optimizer: torch optimizer
+        scheduler: torch scheduler
+        **kwargs: some extra info to pack
+
+    Returns:
+        torch-based checkpoint with ``model_state_dict``,
+        ``criterion_state_dict``, ``optimizer_state_dict``,
+        ``scheduler_state_dict`` keys.
+    """
+    checkpoint = kwargs
+
+    if isinstance(model, dict):
+        for key, value in model.items():
+            model_module = get_nn_from_ddp_module(value)
+            checkpoint[f"model_{key}_state_dict"] = maybe_recursive_call(
+                model_module, "state_dict"
+            )
+    else:
+        model_module = get_nn_from_ddp_module(model)
+        checkpoint["model_state_dict"] = maybe_recursive_call(model_module, "state_dict")
+
+    for dict2save, name2save in zip(
+        [criterion, optimizer, scheduler], ["criterion", "optimizer", "scheduler"],
+    ):
+        if dict2save is None:
+            continue
+        if isinstance(dict2save, dict):
+            for key, value in dict2save.items():
+                if value is not None:
+                    state_dict2save = name2save + "_" + str(key)
+                    # checkpoint[name2save_] = value
+                    state_dict2save = state_dict2save + "_state_dict"
+                    checkpoint[state_dict2save] = value.state_dict()
+        else:
+            # checkpoint[name2save] = dict2save
+            name2save = name2save + "_state_dict"
+            checkpoint[name2save] = dict2save.state_dict()
+    return checkpoint
+
+
+def unpack_checkpoint(
+    checkpoint: Dict,
+    model: RunnerModel = None,
+    criterion: RunnerCriterion = None,
+    optimizer: RunnerOptimizer = None,
+    scheduler: RunnerScheduler = None,
+) -> None:
+    """Load checkpoint from file and unpack the content to a model
+    (if not None), criterion (if not None), optimizer (if not None),
+    scheduler (if not None).
+
+    Args:
+        checkpoint: checkpoint to load
+        model: model where should be updated state
+        criterion: criterion where should be updated state
+        optimizer: optimizer where should be updated state
+        scheduler: scheduler where should be updated state
+    """
+    if model is not None:
+        model = get_nn_from_ddp_module(model)
+        maybe_recursive_call(
+            model, "load_state_dict", recursive_args=checkpoint["model_state_dict"],
+        )
+
+    for dict2load, name2load in zip(
+        [criterion, optimizer, scheduler], ["criterion", "optimizer", "scheduler"],
+    ):
+        if dict2load is None:
+            continue
+
+        if isinstance(dict2load, dict):
+            for key, value in dict2load.items():
+                if value is not None:
+                    state_dict2load = f"{name2load}_{key}_state_dict"
+                    value.load_state_dict(checkpoint[state_dict2load])
+        else:
+            name2load = f"{name2load}_state_dict"
+            dict2load.load_state_dict(checkpoint[name2load])
+
+
+def save_checkpoint(checkpoint: Mapping[str, Any], path: str):
+    """Saves checkpoint to a file.
+
+    Args:
+        checkpoint: data to save.
+        path: filepath where checkpoint should be stored.
+    """
+    torch.save(checkpoint, path)
+
+
+def load_checkpoint(path: str):
+    """Load checkpoint from path.
+
+    Args:
+        path: checkpoint file to load
+
+    Returns:
+        checkpoint content
+    """
+    checkpoint = torch.load(path, map_location=lambda storage, loc: storage)
+    return checkpoint
+
+
 __all__ = [
     "get_optimizable_params",
     "get_optimizer_momentum",
@@ -444,9 +572,13 @@ __all__ = [
     "get_requires_grad",
     "set_requires_grad",
     "get_network_output",
-    "detach",
+    "detach_tensor",
     "trim_tensors",
     "get_optimal_inner_init",
     "outer_init",
     "reset_weights_if_possible",
+    "pack_checkpoint",
+    "unpack_checkpoint",
+    "save_checkpoint",
+    "load_checkpoint",
 ]
